Add download script for daily saves
It downloads the json for Zurich, Budapest and Krakow.
Saves the raw json to daily-saves (in gitignore).

Ignores what cities are actually in the DB.
Used for safekeeping. The minimal amount of logic is
added here so that it never fails and so we at least
always have the daily weather forecast saved.

To be added to crontab